+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 5.1.1 - 2026-03-17
 
 ### Fixed
 - Fixed a bug where Categories (Multiple Groups) fields weren't having their sources setting saved